+++ samanthas_telegram_bot/api_clients/__init__.py
@@ -1,5 +1,5 @@
-from samanthas_telegram_bot.api_clients.backend.backend_client import BackendClient
-from samanthas_telegram_bot.api_clients.smalltalk.smalltalk_client import SmallTalkClient
+from samanthas_telegram_bot.api_clients.backend.client import BackendClient
+from samanthas_telegram_bot.api_clients.smalltalk.client import SmallTalkClient
 
 __all__ = [
     "BackendClient",
