+++ samanthas_telegram_bot/conversation/callbacks/registration/student.py
@@ -1,9 +1,8 @@
 from telegram import Update
 from telegram.ext import ConversationHandler
 
-from samanthas_telegram_bot.api_clients import BackendClient
+from samanthas_telegram_bot.api_clients import BackendClient, SmallTalkClient
 from samanthas_telegram_bot.api_clients.smalltalk.exceptions import SmallTalkClientError
-from samanthas_telegram_bot.api_clients.smalltalk.smalltalk_client import SmallTalkClient
 from samanthas_telegram_bot.auxil.log_and_notify import logs
 from samanthas_telegram_bot.conversation.auxil.callback_query_reply_sender import (
     CallbackQueryReplySender as CQReplySender,
